Ported to double

+++ examples/md-flexible/src/TypeDefinitions.h
@@ -43,7 +43,7 @@
 /**
  * Precision used for particle representations. If you want to test other precisions change it here.
  */
-using FloatPrecision = float;
+using FloatPrecision = double;
 
 /**
  * Type of the Particles used in md-flexible.
